requirements.txt
Django==3.1.7
beautifulsoup4==4.9.3
celery==5.1.2
django-ckeditor==6.1.0
Pillow==8.3.1
redis
gunicorn==20.1.0
psycopg2-binary==2.9.1
celery==5.1.2
django-mptt==0.13.2
openpyxl
